requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.25.0
python-dotenv>=1.0.0
pandas>=1.5.0
langchain-google-genai
google-generativeai
termcolor>=2.4.0
pyngrok>=7.0.0
SQLAlchemy>=2.0.0
